API improvements: unify ifindex to u32, add helper methods
- Unify all public ifindex types to u32 (was mixed i32/u32)
  - LinkMessage::ifindex() -> u32
  - AddressMessage::ifindex() -> u32
  - NeighborMessage::ifindex() -> u32
  - TcMessage::ifindex() -> u32
  - NetworkEvent::ifindex() -> Option<u32>
  - get_ifindex/get_ifindex_opt return u32

- Add LinkMessage::name_or(default) helper method
  - Replaces verbose .name.as_deref().unwrap_or("?") pattern

- Add Connection::get_interface_names() helper
  - Returns HashMap<u32, String> for resolving ifindex to names
  - Useful when displaying routes, addresses, neighbors

- Update all examples to use new helpers
- Add comprehensive examples/README.md
- Update main README.md and CLAUDE.md with new patterns

+++ bins/ip/src/commands/monitor.rs
@@ -149,7 +149,7 @@ fn parse_ip_event(
                         action,
                         address: address.to_string(),
                         prefix_len: addr.prefix_len(),
-                        ifindex: addr.ifindex() as i32,
+                        ifindex: addr.ifindex(),
                         family: addr.family(),
                         scope: addr.scope().name(),
                         label: addr.label.clone(),
@@ -189,7 +189,7 @@ fn parse_ip_event(
                         action,
                         destination: dst.to_string(),
                         lladdr: neigh.mac_address(),
-                        ifindex: neigh.ifindex() as i32,
+                        ifindex: neigh.ifindex(),
                         state: nud_state_name(neigh.header.ndm_state),
                         router: neigh.is_router(),
                     })
